+++ dummyBackend/package.json
@@ -2,15 +2,18 @@
   "name": "dummyBackend",
   "version": "1.0.0",
   "description": "",
-  "main": "api/index.js",
+  "main": "index.js",
   "scripts": {
-    "start": "node api/index.js"
+    "dev": "NODE_ENV=development node index.js",
+    "start": "NODE_ENV=production node index.js",
+    "watch": "json-server db.json --port=8000"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "clone": "^2.1.2",
     "json-serve": "^0.1.0",
     "json-server": "^0.17.3"
   }
