--- dummyBackend/vercel.json
@@ -1,14 +1,0 @@
-{
-    "functions": {
-      "api/index.js": {
-        "memory": 1024,
-        "includeFiles": "db.json"
-      }
-    },
-    "rewrites": [
-      {
-        "source": "/(.*)",
-        "destination": "api/index.js"
-      }
-    ]
-  }
